rename file, remove connect function, clientCreatrion error return added
