+++ rpc/legacy_stream/stream.go
@@ -1,4 +1,4 @@
-package rpc
+package legacy_stream
 
 type Stream interface {
 	Send()
